# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "AdminUser": {
        "additionalProperties": false,
        "properties": {
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "displayName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Displayname"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "isActive": {
            "title": "Isactive",
            "type": "boolean"
          },
          "lastLoginAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lastloginat"
          },
          "role": {
            "enum": [
              "admin",
              "user",
              "tester",
              "guest"
            ],
            "title": "Role",
            "type": "string"
          },
          "username": {
            "title": "Username",
            "type": "string"
          }
        },
        "required": [
          "id",
          "username",
          "email",
          "displayName",
          "role",
          "isActive",
          "createdAt",
          "lastLoginAt"
        ],
        "title": "AdminUser",
        "type": "object"
      },
      "AdminUserPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AdminUser"
            },
            "title": "Items",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "title": "AdminUserPage",
        "type": "object"
      },
      "AttemptItem": {
        "additionalProperties": false,
        "properties": {
          "event": {
            "discriminator": {
              "mapping": {
                "element-name": "#/components/schemas/ElementNameAttempt",
                "nomenclature": "#/components/schemas/NomenclatureAttempt",
                "periodic-table": "#/components/schemas/PeriodicTableAttempt"
              },
              "propertyName": "mode"
            },
            "oneOf": [
              {
                "$ref": "#/components/schemas/ElementNameAttempt"
              },
              {
                "$ref": "#/components/schemas/PeriodicTableAttempt"
              },
              {
                "$ref": "#/components/schemas/NomenclatureAttempt"
              }
            ],
            "title": "Event"
          },
          "receivedAt": {
            "format": "date-time",
            "title": "Receivedat",
            "type": "string"
          },
          "serverSeq": {
            "title": "Serverseq",
            "type": "integer"
          }
        },
        "required": [
          "event",
          "serverSeq",
          "receivedAt"
        ],
        "title": "AttemptItem",
        "type": "object"
      },
      "AttemptPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AttemptItem"
            },
            "title": "Items",
            "type": "array"
          },
          "nextCursor": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Nextcursor"
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "title": "AttemptPage",
        "type": "object"
      },
      "AttemptStats": {
        "additionalProperties": false,
        "properties": {
          "byMode": {
            "items": {
              "$ref": "#/components/schemas/ModeStats"
            },
            "title": "Bymode",
            "type": "array"
          },
          "correctAttempts": {
            "title": "Correctattempts",
            "type": "integer"
          },
          "totalAttempts": {
            "title": "Totalattempts",
            "type": "integer"
          }
        },
        "required": [
          "totalAttempts",
          "correctAttempts",
          "byMode"
        ],
        "title": "AttemptStats",
        "type": "object"
      },
      "BatchRequest": {
        "additionalProperties": false,
        "properties": {
          "events": {
            "items": {
              "discriminator": {
                "mapping": {
                  "element-name": "#/components/schemas/ElementNameAttempt",
                  "nomenclature": "#/components/schemas/NomenclatureAttempt",
                  "periodic-table": "#/components/schemas/PeriodicTableAttempt"
                },
                "propertyName": "mode"
              },
              "oneOf": [
                {
                  "$ref": "#/components/schemas/ElementNameAttempt"
                },
                {
                  "$ref": "#/components/schemas/PeriodicTableAttempt"
                },
                {
                  "$ref": "#/components/schemas/NomenclatureAttempt"
                }
              ]
            },
            "maxItems": 200,
            "minItems": 1,
            "title": "Events",
            "type": "array"
          }
        },
        "required": [
          "events"
        ],
        "title": "BatchRequest",
        "type": "object"
      },
      "BatchResponse": {
        "additionalProperties": false,
        "properties": {
          "accepted": {
            "items": {
              "type": "string"
            },
            "title": "Accepted",
            "type": "array"
          },
          "duplicates": {
            "items": {
              "type": "string"
            },
            "title": "Duplicates",
            "type": "array"
          }
        },
        "required": [
          "accepted",
          "duplicates"
        ],
        "title": "BatchResponse",
        "type": "object"
      },
...
      "DailyTrend": {
        "additionalProperties": false,
        "properties": {
          "correctAttempts": {
            "title": "Correctattempts",
            "type": "integer"
          },
          "day": {
            "title": "Day",
            "type": "string"
          },
          "totalAttempts": {
            "title": "Totalattempts",
            "type": "integer"
          }
        },
        "required": [
          "day",
          "totalAttempts",
          "correctAttempts"
        ],
        "title": "DailyTrend",
        "type": "object"
      },
      "ElementNameAttempt": {
        "additionalProperties": false,
        "properties": {
          "contentVersion": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Contentversion",
            "type": "string"
          },
          "direction": {
            "const": "symbol-to-name",
            "title": "Direction",
            "type": "string"
          },
          "id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "isCorrect": {
            "title": "Iscorrect",
            "type": "boolean"
          },
          "matchPolicy": {
            "const": "diacritics-tolerant",
            "title": "Matchpolicy",
            "type": "string"
          },
          "mode": {
            "const": "element-name",
            "title": "Mode",
            "type": "string"
          },
          "occurredAt": {
            "format": "date-time",
            "title": "Occurredat",
            "type": "string"
          },
          "questionId": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Questionid",
            "type": "string"
          },
          "round": {
            "enum": [
              "initial",
              "retry"
            ],
            "title": "Round",
            "type": "string"
          }
        },
        "required": [
          "id",
          "questionId",
          "contentVersion",
          "occurredAt",
          "isCorrect",
          "round",
          "mode",
          "direction",
          "matchPolicy"
        ],
        "title": "ElementNameAttempt",
        "type": "object"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "additionalProperties": true,
            "title": "Details",
            "type": "object"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "requestId": {
            "title": "Requestid",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message",
          "requestId"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "ModeStats": {
        "additionalProperties": false,
        "properties": {
          "correctAttempts": {
            "title": "Correctattempts",
            "type": "integer"
          },
          "mode": {
            "enum": [
              "element-name",
              "periodic-table",
              "nomenclature"
            ],
            "title": "Mode",
            "type": "string"
          },
          "totalAttempts": {
            "title": "Totalattempts",
            "type": "integer"
          }
        },
        "required": [
          "mode",
          "totalAttempts",
          "correctAttempts"
        ],
        "title": "ModeStats",
        "type": "object"
      },
...
      "PeriodicTableAttempt": {
        "additionalProperties": false,
        "properties": {
          "contentVersion": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Contentversion",
            "type": "string"
          },
          "direction": {
            "enum": [
              "name-to-position",
              "position-to-name",
              "position-to-name-or-symbol",
              "name-to-symbol",
              "symbol-to-name"
            ],
            "title": "Direction",
            "type": "string"
          },
          "id": {
            "maxLength": 128,
            "minLength": 1,
            "title": "Id",
            "type": "string"
          },
          "isCorrect": {
            "title": "Iscorrect",
            "type": "boolean"
          },
          "matchPolicy": {
            "enum": [
              "exact-position",
              "diacritics-tolerant",
              "name-tolerant-or-symbol-exact",
              "symbol-exact"
            ],
            "title": "Matchpolicy",
            "type": "string"
          },
          "mode": {
            "const": "periodic-table",
            "title": "Mode",
            "type": "string"
          },
          "occurredAt": {
            "format": "date-time",
            "title": "Occurredat",
            "type": "string"
          },
          "questionId": {
            "maxLength": 256,
            "minLength": 1,
            "title": "Questionid",
            "type": "string"
          },
          "round": {
            "enum": [
              "initial",
              "retry"
            ],
            "title": "Round",
            "type": "string"
          }
        },
        "required": [
          "id",
          "questionId",
          "contentVersion",
          "occurredAt",
          "isCorrect",
          "round",
          "mode",
          "direction",
          "matchPolicy"
        ],
        "title": "PeriodicTableAttempt",
        "type": "object"
      },
...
      "Progression": {
        "additionalProperties": false,
        "properties": {
          "accuracy": {
            "title": "Accuracy",
            "type": "number"
          },
          "correctAttempts": {
            "title": "Correctattempts",
            "type": "integer"
          },
          "rank": {
            "$ref": "#/components/schemas/Rank"
          },
          "totalAttempts": {
            "title": "Totalattempts",
            "type": "integer"
          },
          "trend": {
            "items": {
              "$ref": "#/components/schemas/DailyTrend"
            },
            "title": "Trend",
            "type": "array"
          }
        },
        "required": [
          "totalAttempts",
          "correctAttempts",
          "accuracy",
          "rank",
          "trend"
        ],
        "title": "Progression",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "InorganicSessionCookie": {
        "description": "Opaque server-side browser session cookie.",
        "in": "cookie",
        "name": "__Host-inorganic_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Progress synchronization and server-owned state for the learning application.",
    "title": "Inorganic Chemistry Learning API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/admin/stats": {
      "get": {
        "operationId": "getAdminStats",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptStats"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "InorganicSessionCookie": []
          }
        ],
        "summary": "Admin Stats",
        "tags": [
          "attempts"
        ]
      }
    },
    "/api/v1/admin/users": {
      "get": {
        "operationId": "listAdminUsers",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminUserPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "InorganicSessionCookie": []
          }
        ],
        "summary": "List Admin Users",
        "tags": [
          "attempts"
        ]
      }
    },
    "/api/v1/admin/users/{user_id}/attempt-events": {
      "get": {
        "operationId": "listUserAttemptEvents",
        "parameters": [
          {
            "in": "path",
            "name": "user_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "User Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "InorganicSessionCookie": []
          }
        ],
        "summary": "List User Attempt Events",
        "tags": [
          "attempts"
        ]
      }
    },
...
    "/api/v1/me/attempt-events": {
      "get": {
        "operationId": "listMyAttemptEvents",
        "parameters": [
          {
            "in": "query",
            "name": "cursor",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 128,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cursor"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "InorganicSessionCookie": []
          }
        ],
        "summary": "List My Attempt Events",
        "tags": [
          "attempts"
        ]
      }
    },
    "/api/v1/me/attempt-events/batch": {
      "post": {
        "operationId": "uploadAttemptEvents",
        "parameters": [
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Csrf-Token"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "InorganicSessionCookie": []
          }
        ],
        "summary": "Upload Attempt Events",
        "tags": [
          "attempts"
        ]
      }
    },
...
    "/api/v1/me/progression": {
      "get": {
        "operationId": "getMyProgression",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Progression"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Forbidden"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "InorganicSessionCookie": []
          }
        ],
        "summary": "My Progression",
        "tags": [
          "attempts"
        ]
      }
    },
    "/api/v1/me/stats": {
      "get": {
        "operationId": "getMyStats",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AttemptStats"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Bad Request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unauthorized"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unprocessable Entity"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Internal Server Error"
          }
        },
        "security": [
          {
            "InorganicSessionCookie": []
          }
        ],
        "summary": "My Stats",
        "tags": [
          "attempts"
        ]
      }
    }
  }
}